Fix up script to generate example
