+++ package.json
@@ -15,11 +15,11 @@
   },
   "scripts": {
     "build": "babel src -d generators",
-    "post-build": "onchange 'generators/**' 'test/**.js' -- npm test",
+    "post-build": "onchange 'generators/**' 'test/**.js' -- npm test && npm run example",
     "watch": "npm run build -- -w & npm run post-build",
     "test": "standard && mocha --compilers js:babel/register",
-    "example": "node generators/example.js",
-    "prepublish": "npm run build && npm test"
+    "example": "node generators/generate-example.js",
+    "prepublish": "npm run build && npm test && npm run example"
   },
   "files": [
     "generators"
@@ -38,7 +38,8 @@
     "mocha": "*",
     "onchange": "1.x",
     "silent-yeoman": "0.0.4",
-    "standard": "*"
+    "standard": "*",
+    "yeoman-environment": "^1.2.5"
   },
   "standard": {
     "ignore": [
